pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "azure-foundry-voice-agent-demo"
version = "0.1.0"
description = "Demo: Azure AI Foundry Voice Agent with Voice Live API and Foundry Agent Service"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "azure-ai-agents>=1.0.0",
    "azure-ai-voicelive[aiohttp]>=1.0.0",
    "azure-identity>=1.15.0",
    "websockets>=12.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py310"
line-length = 100
